requirements.txt
# Multilingual RAG Voice Assistant (LiveKit agent + token server)
# Python 3.10+ recommended. Install: pip install -r requirements.txt

# --- LiveKit (livekit_agent.py, token_server.py) ---
livekit==1.1.3
livekit-api==1.1.0
livekit-protocol==1.1.3
fastapi==0.135.2
uvicorn==0.42.0
starlette==1.0.0
PyJWT==2.12.1
types-protobuf==6.32.1.20260221

# --- Audio I/O & STT/TTS ---
av==17.0.0
faster-whisper==1.2.1
ctranslate2==4.7.1
onnxruntime==1.24.4
edge-tts==7.2.7

# --- RAG / embeddings ---
numpy==2.4.3
sentence-transformers==5.3.0
scikit-learn==1.8.0
scipy==1.17.1
torch==2.10.0
transformers==5.3.0
tokenizers==0.22.2
safetensors==0.7.0
huggingface_hub==1.7.1
hf-xet==1.4.2

# --- LLM & translation ---
groq==1.1.1
deep-translator==1.11.4
requests==2.32.5

# --- Config & HTTP ---
python-dotenv==1.2.2
aiohttp==3.13.3
httpx==0.28.1
httpcore==1.0.9

# --- Shared utilities (pulled by deps above; pinned for reproducibility) ---
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
filelock==3.25.2
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
packaging==26.0
propcache==0.4.1
protobuf==7.34.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyYAML==6.0.3
regex==2026.2.28
rich==14.3.3
setuptools==82.0.1
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.8.3
sympy==1.14.0
tabulate==0.10.0
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yarl==1.23.0
